# packages/yucca-admin-api/openapi-specs.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/user": {
      "get": {
        "operationId": "listUsers",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 50,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/api/user/{id}": {
      "get": {
        "operationId": "getUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGetResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ]
      },
      "patch": {
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserUpdateResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ]
      },
      "delete": {
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/api/user/{userId}/session": {
      "get": {
        "operationId": "listUserSessions",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ]
      },
      "delete": {
        "operationId": "deleteUserSessions",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/api/user/{id}/features": {
      "get": {
        "operationId": "getUserFeatures",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserFeaturesResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/api/user/{id}/features/{flag}": {
      "put": {
        "operationId": "setUserFeature",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "flag",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeatureOverrideSetRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureOverrideDto"
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ]
      },
      "delete": {
        "operationId": "clearUserFeature",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "flag",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/api/user/{id}/consumers": {
      "get": {
        "operationId": "listUserConsumers",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsumerListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ]
      }
    },
...
    "/api/repository": {
      "get": {
        "operationId": "listRepositories",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 50,
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Repository"
        ]
      },
      "post": {
        "operationId": "createRepository",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepositoryCreateRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryCreateResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Repository"
        ]
      }
    },
    "/api/repository/{id}": {
      "get": {
        "operationId": "getRepository",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryGetResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Repository"
        ]
      },
      "patch": {
        "operationId": "updateRepository",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepositoryUpdateRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryUpdateResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Repository"
        ]
      },
      "delete": {
        "operationId": "deleteRepository",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Repository"
        ]
      }
    },
...
    "/api/restic-tokens": {
      "get": {
        "operationId": "listResticTokens",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 50,
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repositoryId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "active",
            "required": false,
            "in": "query",
            "description": "true = only unexpired, unrevoked tokens",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResticTokenListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "ResticToken"
        ]
      }
    },
    "/api/restic-tokens/{jti}": {
      "delete": {
        "operationId": "revokeResticToken",
        "parameters": [
          {
            "name": "jti",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "ResticToken"
        ]
      }
    },
...
  },
  "info": {
    "title": "yucca-admin",
    "description": "yucca admin API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "schemas": {
...
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "sub": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "disabled": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "sub",
          "name",
          "email",
          "disabled"
        ]
      },
      "UserListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserDto"
            }
          },
          "nextCursor": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "items"
        ]
      },
      "UserGetResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserDto"
          }
        },
        "required": [
          "user"
        ]
      },
      "UserUpdateRequestDto": {
        "type": "object",
        "properties": {
          "disabled": {
            "type": "boolean"
          }
        }
      },
      "UserUpdateResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserDto"
          }
        },
        "required": [
          "user"
        ]
      },
      "SessionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "userId"
        ]
      },
      "SessionListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SessionDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
      "FeatureOverrideDto": {
        "type": "object",
        "properties": {
          "flag": {
            "type": "string"
          },
          "value": {
            "type": "boolean"
          },
          "setBy": {
            "type": "string"
          },
          "reason": {
            "type": "object",
            "nullable": true
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "flag",
          "value",
          "setBy",
          "updatedAt"
        ]
      },
      "UserFeaturesResponseDto": {
        "type": "object",
        "properties": {
          "features": {
            "type": "object",
            "additionalProperties": {
              "type": "boolean"
            },
            "description": "Resolved flags: override, else registry default"
          },
          "overrides": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FeatureOverrideDto"
            }
          }
        },
        "required": [
          "features",
          "overrides"
        ]
      },
      "FeatureOverrideSetRequestDto": {
        "type": "object",
        "properties": {
          "value": {
            "type": "boolean"
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "value"
        ]
      },
      "ConsumerAdminDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "description": "immich | fubar | restic"
          },
          "name": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "lastSeenAt": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "userId",
          "type",
          "name",
          "createdAt"
        ]
      },
      "ConsumerListResponseDto": {
        "type": "object",
        "properties": {
          "consumers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConsumerAdminDto"
            }
          }
        },
        "required": [
          "consumers"
        ]
      },
      "RepositoryOwnerDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "disabled": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "disabled"
        ]
      },
      "RepositoryMetricsDto": {
        "type": "object",
        "properties": {
          "sizeBytes": {
            "type": "number"
          },
          "lastStarted": {
            "type": "string",
            "nullable": true
          },
          "lastBackup": {
            "type": "string",
            "nullable": true
          },
          "lastSuccessfulBackup": {
            "type": "string",
            "nullable": true
          },
          "lastBackupDuration": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "sizeBytes"
        ]
      },
      "RepositoryAdminDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "worm": {
            "type": "boolean"
          },
          "consumerId": {
            "type": "string"
          },
          "consumerType": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/RepositoryOwnerDto"
          },
          "metrics": {
            "$ref": "#/components/schemas/RepositoryMetricsDto"
          }
        },
        "required": [
          "id",
          "name",
          "worm",
          "consumerId",
          "consumerType",
          "user",
          "metrics"
        ]
      },
      "RepositoryListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RepositoryAdminDto"
            }
          },
          "nextCursor": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "items"
        ]
      },
      "RepositoryCreateRequestDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "userId": {
            "type": "string",
            "description": "Owner; defaults to the admin service user"
          },
          "worm": {
            "type": "boolean"
          },
          "consumerType": {
            "type": "string",
            "enum": [
              "immich",
              "fubar",
              "restic"
            ],
            "description": "Consumer type for the owning consumer; defaults to 'restic' (manual use)"
          }
        },
        "required": [
          "name"
        ]
      },
      "RepositoryCreateResponseDto": {
        "type": "object",
        "properties": {
          "repository": {
            "$ref": "#/components/schemas/RepositoryAdminDto"
          }
        },
        "required": [
          "repository"
        ]
      },
      "RepositoryGetResponseDto": {
        "type": "object",
        "properties": {
          "repository": {
            "$ref": "#/components/schemas/RepositoryAdminDto"
          }
        },
        "required": [
          "repository"
        ]
      },
...
      "RepositoryUpdateRequestDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "worm": {
            "type": "boolean"
          }
        }
      },
      "RepositoryUpdateResponseDto": {
        "type": "object",
        "properties": {
          "repository": {
            "$ref": "#/components/schemas/RepositoryAdminDto"
          }
        },
        "required": [
          "repository"
        ]
      },
      "ResticTokenDto": {
        "type": "object",
        "properties": {
          "jti": {
            "type": "string"
          },
          "repositoryId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "consumerId": {
            "type": "object",
            "nullable": true
          },
          "mintedBy": {
            "type": "string",
            "description": "'user' or 'admin'"
          },
          "label": {
            "type": "object",
            "nullable": true
          },
          "expiresAt": {
            "type": "string"
          },
          "revokedAt": {
            "type": "string",
            "nullable": true
          },
          "revokedBy": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "jti",
          "repositoryId",
          "userId",
          "mintedBy",
          "expiresAt",
          "createdAt"
        ]
      },
      "ResticTokenListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResticTokenDto"
            }
          },
          "nextCursor": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "items"
        ]
      },
...
    }
  }
}